requirements.txt
websocket-client
numpy
matplotlib